requirements.txt
python-pptx==0.6.23
requests==2.31.0
python-dotenv==1.0.1
openai==1.14.0
google-generativeai==0.4.1
